refactor(type_checker): move tests into tests.rs
